requirements.txt
pyzabbix==1.3.1
python-dotenv==1.0.0
schedule==1.2.0
